+++ web/lib/container_view.dart
@@ -13,35 +13,33 @@ abstract class ContainerView {
 
   LinkElement styleLink;
   AnchorElement tabHandleButton;
-  DivElement content;
+  DivElement content, tabContainer, tabContent;
   LIElement tabHandle;
   UListElement menus;
 
-  DivElement _tabContainer, _tabContent;
-
-  ContainerView(this.id, this.col, this.title, this.shortName, this.config) {
+  ContainerView(this.id, this.col, this.title, this.shortName, this.config, DivElement handles) {
     refMap = {};
 
-    _setUpTabHandle();
+    _setUpTabHandle(handles);
     _setUpTabContainer();
   }
 
   /// Adds the CSS classes to make a tab 'active'.
   void makeActive() {
     tabHandle.classes.add('active');
-    _tabContainer.classes.add('active');
+    tabContainer.classes.add('active');
   }
 
   /// Removes the CSS classes to make a tab 'inactive'.
   void makeInactive() {
     tabHandle.classes.remove('active');
-    _tabContainer.classes.remove('active');
+    tabContainer.classes.remove('active');
   }
 
   /// Removes the tab elements from the DOM.
   void destroy() {
     tabHandle.remove();
-    _tabContainer.remove();
+    tabContainer.remove();
     if (styleLink != null) styleLink.remove();
   }
 
@@ -83,7 +81,7 @@ abstract class ContainerView {
   }
 
   /// Takes a [num], [col], and [title] to add a new tab for the specified column.
-  void _setUpTabHandle() {
+  void _setUpTabHandle(DivElement handles) {
     tabHandle = new LIElement()
       ..classes.add('tab-handle')
       ..classes.add('active');
@@ -96,8 +94,7 @@ abstract class ContainerView {
         ..dataset['toggle'] = 'tab';
     tabHandle.children.add(tabHandleButton);
 
-    DivElement column = querySelector('#column-$col');
-    column.children.first.children.add(tabHandle);
+    handles.children.add(tabHandle);
   }
 
   /// Takes a [num], [col], [title], [config], and [active] to generate the menu bar and menu items
@@ -105,7 +102,7 @@ abstract class ContainerView {
   void _setUpTabContainer() {
     String name = title.toLowerCase().replaceAll(' ', '-');
 
-    _tabContainer = new DivElement()
+    tabContainer = new DivElement()
         ..id = 'tab-$name-$id-container'
         ..classes.add('tab-pane')
         ..classes.add('active');
@@ -115,24 +112,25 @@ abstract class ContainerView {
         ..classes.add('nav-tabs')
         ..classes.add('inner-tabs')
         ..attributes['role'] = 'tablist';
-    _tabContainer.children.add(menus);
+    tabContainer.children.add(menus);
 
     menus.children = new List<Element>();
     for (Map configItem in config) {
       menus.children.add(_createDropdownMenu(configItem));
     }
 
-    _tabContent = new DivElement()
-        ..classes.add('tab-content');
-    _tabContainer.children.add(_tabContent);
+    tabContent = new DivElement()
+        ..classes.add('tab-content')
+        ..tabIndex = -1;
+    tabContainer.children.add(tabContent);
 
     content = new DivElement()
         ..classes.add(name);
-    _tabContent.children.add(content);
+    tabContent.children.add(content);
     refMap['content'] = content;
 
     DivElement colOneTabContent = querySelector('#col-$col-tab-content');
-    colOneTabContent.children.insert(0, _tabContainer);
+    colOneTabContent.children.insert(0, tabContainer);
   }
 
   /// Generates a dropdown menu and returns the new [LIElement].