+++ web/lib/modal/open_tab.dart
@@ -1,11 +1,9 @@
 part of updroid_modal;
 
 class UpDroidOpenTabModal extends UpDroidModal {
-  int _side;
   Function _openTab;
 
-  UpDroidOpenTabModal(int side, Function openTab) {
-    _side = side;
+  UpDroidOpenTabModal(Function openTab) {
     _openTab = openTab;
 
     _setupHead('Select Tab: ');
@@ -20,11 +18,11 @@ class UpDroidOpenTabModal extends UpDroidModal {
       ..id = "selector-wrapper";
     _modalBody.children.add(selectorWrap);
 
-    ButtonElement sEditor = _createButton('default', 'Editor')..onClick.listen((_) => _openTab(_side, 'UpDroidEditor'));
-    ButtonElement sConsole = _createButton('default', 'Console')..onClick.listen((_) => _openTab(_side, 'UpDroidConsole'));
-    ButtonElement sCamera = _createButton('default', 'Camera')..onClick.listen((_) => _openTab(_side, 'UpDroidCamera'));
-    ButtonElement sTeleop = _createButton('default', 'Teleop')..onClick.listen((_) => _openTab(_side, 'UpDroidTeleop'));
-    selectorWrap.children.addAll([sEditor, sConsole, sCamera]);
+    ButtonElement sEditor = _createButton('default', 'Editor')..onClick.listen((_) => _openTab('UpDroidEditor'));
+    ButtonElement sConsole = _createButton('default', 'Console')..onClick.listen((_) => _openTab('UpDroidConsole'));
+    ButtonElement sCamera = _createButton('default', 'Camera')..onClick.listen((_) => _openTab('UpDroidCamera'));
+    ButtonElement sTeleop = _createButton('default', 'Teleop')..onClick.listen((_) => _openTab('UpDroidTeleop'));
+    selectorWrap.children.addAll([sEditor, sConsole, sCamera, sTeleop]);
   }
 
   void _setupFooter() {