+++ web/lib/panels/panel_view.dart
@@ -14,7 +14,7 @@ class PanelView extends ContainerView {
   }
 
   PanelView(int id, int col, String title, String shortName, List config, [bool externalCss=false]) :
-  super(id, col, title, shortName, config) {
+  super(id, col, title, shortName, config, querySelector('#column-$col').children[1]) {
     if (externalCss) {
       String cssPath = 'lib/panels/${shortName.toLowerCase()}/${shortName.toLowerCase()}.css';
       loadExternalCss(cssPath);